requirements.txt
pytest
playwright
pytest-playwright
python-dotenv
allure-pytest